% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{aicc}
\alias{aicc}
\title{Corrected AIC}
\usage{
aicc(x, n = Inf)
}
\arguments{
\item{x}{maxLik object}

\item{n}{number of observations for correction. Defaults to Inf, which coincides with the AIC.}
}
\value{
numeric
}
\description{
Compute the corrected AIC from a maxLik object.
}
\keyword{AICc}

