% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{areFixed}
\alias{areFixed}
\title{Mainly auxiliary. Returns a vector indicating which parameters are fixed at zero.}
\usage{
areFixed(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Returns a vector indicating which parameters are fixed at zero.
}
\keyword{fixed}
\keyword{parameters}

