% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{as.numeric.matrix}
\alias{as.numeric.matrix}
\title{Mainly auxiliary. Same as as.numeric() but for matrices.}
\usage{
\method{as.numeric}{matrix}(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
matrix
}
\description{
Function to return a matrix with numericals.
}
\keyword{as}
\keyword{numeric}

