% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{breads}
\alias{breads}
\title{Get the breads to build standard errors.}
\usage{
breads(model)
}
\arguments{
\item{model}{maxLik object}
}
\value{
matrix
}
\description{
Computes the inverse Fischer Information matrix
}
\keyword{Information}
\keyword{Inverse}
\keyword{matrix}

