% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{dropna}
\alias{dropna}
\title{Drop NA values from a vector.}
\usage{
dropna(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Removes NA values from a vector.
}
\keyword{drop}
\keyword{missing}
\keyword{values}

