% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{freePars}
\alias{freePars}
\title{Get number of unrestricted parameters from a maxLik object.}
\usage{
freePars(x)
}
\arguments{
\item{x}{maxLik object,}
}
\value{
numeric
}
\description{
Get number of unrestricted parameters from a maxLik object.
}
\keyword{free}
\keyword{parameters}

