% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{parsToOriginal2}
\alias{parsToOriginal2}
\title{Mainly auxiliary. Function to convert final parameters back into the format of the starting parameter vectors.}
\usage{
parsToOriginal2(results, start)
}
\arguments{
\item{results}{vector of parameters}

\item{start}{vector of the original starting values.}
}
\value{
vector
}
\description{
Function to convert final parameters back into the format of the starting parameter vectors. Sets parameters that are contstrained in the final results to supplied starting values.
}
\keyword{estimates.}
\keyword{parameter}
\keyword{reformat}

