% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optMaxlik.R
\name{se}
\alias{se}
\title{Get standard errors.}
\usage{
se(model)
}
\arguments{
\item{model}{maxLik object}
}
\value{
matrix
}
\description{
Computes the standard errors from the Hessian.
}
\keyword{errors}
\keyword{standard}

